添加接近光驱动

+++ model/sensor/driver/proximity/sensor_proximity_driver.h
@@ -18,7 +18,7 @@
 #define PROXIMITY_CHIP_NAME_APDS9960        "apds9960"
 
 struct ProximityData {
-    uint32_t stateFlag;
+    uint8_t stateFlag;
 };
 
 struct ProximityOpsCall {
